Moved canvas functionality into its own component.
